pyproject.toml
[project]
name = "pyline"
version = "0.1.0"
description = "A Sonolus Rizline Engine"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "sonolus-py>=0.13.2",
]
